pyproject.toml
[project]
name = "dataset"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "lightgbm>=4.6.0",
    "lxml>=6.0.2",
    "numpy>=2.4.1",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "tqdm>=4.67.3",
]
